9ns: Add 'extensions' to dev.create
This is based on Eric VH's 9p2000 RFC:

	http://ericvh.github.io/9p-rfc/rfc9p2000.u.html

The extension string's meaning will depend on the mode/perm bits (not
omode, which is open mode).

The main user will be symlinks.  The ext string will be the path.  By
adding the path to the create message, we can atomically create a symlink
file and have its path set.  That might make namec easier.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>